feat: error code on streaming response throw/fail
Updates the `write-error!` method in streaming-response to check if the
current Servlet Request has been 'committed' -- started writing to the
client -- and if it has not, to set a new status code and content type
on the response. This works because the HTTP header is not written until
the first byte is written to the async response's output stream.

This doesn't catch __all__ errors that could happen from the qp, if we
have started writing a response and some row a thousand lines in throws
we will still return a 200 and somewhere in the body response there may
be a mangled error message, but most errors in qp response seem to be
'we cannot run the query', so this will at least provide that error
clearly.

It also seems the download status widget is already setup to work with
this error message.